pyproject.toml
[project]
name = "products-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.128.0",
    "greenlet>=3.3.0",
    "psycopg2>=2.9.11",
    "pydantic>=2.12.5",
    "pydotenv>=0.0.7",
    "sqlmodel>=0.0.31",
]
